+++ src/lib/core/base/base2.ts
@@ -20,6 +20,7 @@ export abstract class BaseDirective2 implements OnChanges, OnDestroy {
   /** The most recently used styles for the builder */
   protected mru: StyleDefinition = {};
   protected destroySubject: Subject<void> = new Subject();
+  protected currentValue: any;
 
   /** Access to host element's parent DOM node */
   protected get parentElement(): HTMLElement | null {
@@ -132,6 +133,10 @@ export abstract class BaseDirective2 implements OnChanges, OnDestroy {
     return 'row';
   }
 
+  protected hasWrap(target: HTMLElement): boolean {
+    return this.styler.hasWrap(target);
+  }
+
   /** Applies styles given via string pair or object map to the directive element */
   protected applyStyleToElement(style: StyleDefinition,
                                 value?: string | number,
@@ -144,6 +149,9 @@ export abstract class BaseDirective2 implements OnChanges, OnDestroy {
   }
 
   protected updateWithValue(input: string) {
-    this.addStyles(input);
+    if (this.currentValue !== input) {
+      this.addStyles(input);
+      this.currentValue = input;
+    }
   }
 }