+++ src/lib/core/breakpoints/data/orientation-break-points.ts
@@ -9,11 +9,11 @@
 import {BreakPoint} from '../break-point';
 
 /* tslint:disable */
-const HANDSET_PORTRAIT  = '(orientation: portrait) and (max-width: 599.99px)';
-const HANDSET_LANDSCAPE = '(orientation: landscape) and (max-width: 959.99px)';
+const HANDSET_PORTRAIT  = '(orientation: portrait) and (max-width: 599.98px)';
+const HANDSET_LANDSCAPE = '(orientation: landscape) and (max-width: 959.98px)';
 
-const TABLET_PORTRAIT   = '(orientation: portrait) and (min-width: 600px) and (max-width: 839.99px)';
-const TABLET_LANDSCAPE  = '(orientation: landscape) and (min-width: 960px) and (max-width: 1279.99px)';
+const TABLET_PORTRAIT   = '(orientation: portrait) and (min-width: 600px) and (max-width: 839.98px)';
+const TABLET_LANDSCAPE  = '(orientation: landscape) and (min-width: 960px) and (max-width: 1279.98px)';
 
 const WEB_PORTRAIT      = '(orientation: portrait) and (min-width: 840px)';
 const WEB_LANDSCAPE     = '(orientation: landscape) and (min-width: 1280px)';
@@ -27,7 +27,7 @@ export const ScreenTypes = {
   'TABLET_PORTRAIT'   : `${TABLET_PORTRAIT} `,
   'WEB_PORTRAIT'      : `${WEB_PORTRAIT}`,
 
-  'HANDSET_LANDSCAPE' : `${HANDSET_LANDSCAPE}]`,
+  'HANDSET_LANDSCAPE' : `${HANDSET_LANDSCAPE}`,
   'TABLET_LANDSCAPE'  : `${TABLET_LANDSCAPE}`,
   'WEB_LANDSCAPE'     : `${WEB_LANDSCAPE}`
 };
@@ -41,7 +41,7 @@ export const ORIENTATION_BREAKPOINTS : BreakPoint[] = [
   {'alias': 'handset.portrait',   priority: 2000, 'mediaQuery': ScreenTypes.HANDSET_PORTRAIT},
 
   {'alias': 'tablet',             priority: 2100, 'mediaQuery': ScreenTypes.TABLET},
-  {'alias': 'tablet.landscape',   priority: 2100, 'mediaQuery': ScreenTypes.TABLET},
+  {'alias': 'tablet.landscape',   priority: 2100, 'mediaQuery': ScreenTypes.TABLET_LANDSCAPE},
   {'alias': 'tablet.portrait',    priority: 2100, 'mediaQuery': ScreenTypes.TABLET_PORTRAIT},
 
   {'alias': 'web',                priority: 2200, 'mediaQuery': ScreenTypes.WEB, overlapping : true },