+++ src/lib/core/media-observer/media-observer.spec.ts
@@ -59,6 +59,7 @@ describe('media-observer', () => {
     it('can supports the `.isActive()` API', () => {
       expect(media$).toBeDefined();
 
+      mediaController.autoRegisterQueries = false;
       // Activate mediaQuery associated with 'md' alias
       mediaController.activate('md');
       expect(mediaObserver.isActive('md')).toBeTruthy();
@@ -126,10 +127,6 @@ describe('media-observer', () => {
 
         activateQuery('gt-lg');
         expect(current.mediaQuery).toEqual(findMediaQuery('gt-lg'));
-
-        activateQuery('unknown');
-        expect(current.mediaQuery).toEqual(findMediaQuery('gt-lg'));
-
       } finally {
         mediaController.autoRegisterQueries = true;
         subscription.unsubscribe();
@@ -181,7 +178,7 @@ describe('media-observer', () => {
   describe('with custom BreakPoints', () => {
     const gtXsMediaQuery = 'screen and (min-width:120px) and (orientation:landscape)';
     const superXLQuery = 'screen and (min-width:10000px)';
-    const smMediaQuery = 'screen and (min-width: 600px) and (max-width: 959.99px)';
+    const smMediaQuery = 'screen and (min-width: 600px) and (max-width: 959.98px)';
 
     const CUSTOM_BREAKPOINTS = [
       {alias: 'slate.xl', priority: 11000, mediaQuery: superXLQuery},
@@ -236,7 +233,7 @@ describe('media-observer', () => {
   });
 
   describe('with layout "print" configured', () => {
-    const mdMediaQuery = 'screen and (min-width: 960px) and (max-width: 1279.99px)';
+    const mdMediaQuery = 'screen and (min-width: 960px) and (max-width: 1279.98px)';
 
     beforeEach(() => {
       // Configure testbed to prepare services
@@ -288,7 +285,7 @@ describe('media-observer', () => {
   });
 
   describe('with layout print NOT configured', () => {
-    const smMediaQuery = 'screen and (min-width: 600px) and (max-width: 959.99px)';
+    const smMediaQuery = 'screen and (min-width: 600px) and (max-width: 959.98px)';
 
     beforeEach(() => {
       // Configure testbed to prepare services