+++ src/lib/flex/flex/flex.ts
@@ -5,7 +5,7 @@
  * Use of this source code is governed by an MIT-style license that can be
  * found in the LICENSE file at https://angular.io/license
  */
-import {Directive, ElementRef, Inject, Injectable, Input} from '@angular/core';
+import {Directive, ElementRef, Inject, Injectable, Input, OnInit} from '@angular/core';
 import {
   BaseDirective2,
   LayoutConfigOptions,
@@ -199,11 +199,12 @@ const selector = `
  *
  * @see https://css-tricks.com/snippets/css/a-guide-to-flexbox/
  */
-export class FlexDirective extends BaseDirective2 {
+@Directive()
+export class FlexDirective extends BaseDirective2 implements OnInit {
 
   protected DIRECTIVE_KEY = 'flex';
-  protected direction = '';
-  protected wrap = false;
+  protected direction?: string = undefined;
+  protected wrap?: boolean = undefined;
 
 
   @Input('fxShrink')
@@ -223,13 +224,16 @@ export class FlexDirective extends BaseDirective2 {
   protected flexGrow = '1';
   protected flexShrink = '1';
 
-  constructor(protected elRef: ElementRef,
-              protected styleUtils: StyleUtils,
+  constructor(elRef: ElementRef,
+              styleUtils: StyleUtils,
               @Inject(LAYOUT_CONFIG) protected layoutConfig: LayoutConfigOptions,
-              protected styleBuilder: FlexStyleBuilder,
+              styleBuilder: FlexStyleBuilder,
               protected marshal: MediaMarshaller) {
     super(elRef, styleBuilder, styleUtils, marshal);
     this.init();
+  }
+
+  ngOnInit() {
     if (this.parentElement) {
       this.marshal.trackValue(this.parentElement, 'layout')
         .pipe(takeUntil(this.destroySubject))
@@ -255,8 +259,11 @@ export class FlexDirective extends BaseDirective2 {
   /** Input to this is exclusively the basis input value */
   protected updateWithValue(value: string) {
     const addFlexToParent = this.layoutConfig.addFlexToParent !== false;
-    if (!this.direction) {
+    if (this.direction === undefined) {
       this.direction = this.getFlexFlowDirection(this.parentElement!, addFlexToParent);
+    }
+    if (this.wrap === undefined) {
+      this.wrap = this.hasWrap(this.parentElement!);
     }
     const direction = this.direction;
     const isHorizontal = direction.startsWith('row');
@@ -279,7 +286,7 @@ export class FlexDirective extends BaseDirective2 {
   protected triggerReflow() {
     const activatedValue = this.activatedValue;
     if (activatedValue !== undefined) {
-      const parts = validateBasis(activatedValue, this.flexGrow, this.flexShrink);
+      const parts = validateBasis(activatedValue + '', this.flexGrow, this.flexShrink);
       this.marshal.updateElement(this.nativeElement, this.DIRECTIVE_KEY, parts.join(' '));
     }
   }