+++ src/lib/grid/row/row.ts
@@ -5,7 +5,7 @@
  * Use of this source code is governed by an MIT-style license that can be
  * found in the LICENSE file at https://angular.io/license
  */
-import {Directive, ElementRef, Optional, Injectable} from '@angular/core';
+import {Directive, ElementRef, Injectable} from '@angular/core';
 import {
   BaseDirective2,
   StyleUtils,
@@ -23,15 +23,14 @@ export class GridRowStyleBuilder extends StyleBuilder {
   }
 }
 
+@Directive()
 export class GridRowDirective extends BaseDirective2 {
   protected DIRECTIVE_KEY = 'grid-row';
 
-  constructor(protected elementRef: ElementRef,
-              // NOTE: not actually optional, but we need to force DI without a
-              // constructor call
-              @Optional() protected styleBuilder: GridRowStyleBuilder,
-              protected styler: StyleUtils,
-              protected marshal: MediaMarshaller) {
+  constructor(elementRef: ElementRef,
+              styleBuilder: GridRowStyleBuilder,
+              styler: StyleUtils,
+              marshal: MediaMarshaller) {
     super(elementRef, styleBuilder, styler, marshal);
     this.init();
   }